#3 Finished frontend file configuration
I have added eslint and typescript as well as confirmed that it works. Upon running npm run dev or npm run build the linting is executed. I have modified the README.md file for the frontend structure to include details and reasoning about how files are structured in the frontend.