#7 Update template for a service
This commit adds a template structure for a service, in this case
the session service. The template uses PostgreSQL as the database
and Sequelize as the ORM. The template also includes a basic
Express server to ensure communication with other services and components.
Lastly, `.gitkeep` is a placeholder file to ensure that the empty
directories are tracked by Git.