#7 Read `db/schema.sql` in db config
This ensures that the schema is loaded into the database when the
service starts. Basically, in case the server is invoked the first
time or for some reason the table is dropped, the schema will be
loaded into the database, hence preventing the service from crashing
due to missing tables.