Fix linting errors
These errors were introduced in the previous commit and detected by
the pipeline that is work in progress. The commit is not associated
with any issue, apart from #12 which introduced these errors. The
problems were minor and easy to fix and the pipeline will not be
blocked by them anymore.