#26 Update `db/config.ts` of `appointment-service`
Previously, the mechanism to create the database file was to check if
the file exists and if not, create it. This was done by using the
`fs.existsSync` function. However, `better-sqlite3` support it out
of the box, only the `fileMustExist` option needs to be set to `false`.
Furthermore, the `verbose` flag is used to log all SQL queries to the
console. This is useful for debugging purposes.