#26 Add scaffolding for `appointment-service` middleware
This commit adds the scaffolding for the `appointment-service` middleware. It
basically adds the different controller function in a simple initial
state as well as the routes to access them. Ultimately, they are
realized in `app.ts`. The routes are defined in `routes/index.ts`.

Postman tests will be added in a later commit.