#14 Refactored code in attempt to help teammate with bugs
I changed the type of the deleteDentist method to allow the returning of status directly to ensure that headers are not sent after they have previously been sent to client. The event listener also needs to be closed to ensure that it does not continue to try to handle messages. I added comments for Maria to check to help her with her bug.