#35 - Added dentist card, formatting, added demo picture, styling
I have finalized the map. I have created helper function that upon invocation, will get all dentists from the dentist API and allow the user find the dentist that they want to visit. Later on, it could be good to perhaps make it so that we get the closest dentists. But this will do for now as an MVP.