#53 #15 - Custom broker, integration tests
I modified the package.json to run the tests while all the services are running. This way we can test the integration of components with each other. In addition, I extracted the broker url to an env file. With the help of Michal Spano, we were able to setup a custom MQTT broker which I also incorporated into the test script.