#46 - Modify CORS Policy in all services
I have modified the CORS policy so that we are allowed by modern browser to pass in httpOnly cookies in corresponding requests. I have done this by creating a function that will check for a certain number of whitelisted origins to see if it is ok to enable origin or not as well as the handling of credentials.