#72 #39 #73 - Added grouping of dentists, connected to backend, slots
I have made it so that dentists are grouped using a third party library that is well respected by the users of leaflet, the library that is used for the map. I have also made it so that a patient can now book an appointment for any date. I have also slightly changed the styling to ensure it aligns with the team's color scheme.