#87 - Large queries, statistics, bug discoveries
In this commit I finalize the chart components, I use a somewhat complex sql query to achieve a filled time series with requests count that provide valuable information to the admin of the website. It includes statistics like: general system usage, number of available appointments, number of registered dentists/patients as well as login requests.