#50 Realize subscription on client
This commit also makes additional adjustments to the client-side code,
namely the `AppointmentsList` component. The component now fetches the
patient's email address and determines if the patient is subscribed to
the selected dentist. This check is carried when the component is
mounted. The status shall only be handled when patient's email is
fetched.

Furthemore, additonal types have been added to this component, and
the patient's email has been extracted to a signal for consistency
and not 'abusing' the API endpoint on the server's side.