#92 - Moved types, notification counter
I have modified the notification counter. Upon logging in and fetching notifications we set an initial value to 0. When we get a new notification this value is updated and the user is informed of a new notification. When you are on the notification page, no notification bells will be shown.